requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
seaborn
Pillow